+++ src/components/Room/Room.tsx
@@ -4,8 +4,9 @@ import { makeStyles, Theme } from '@material-ui/core';
 import ChatWindow from '../ChatWindow/ChatWindow';
 import ParticipantList from '../ParticipantList/ParticipantList';
 import MainParticipant from '../MainParticipant/MainParticipant';
+import BackgroundSelectionDialog from '../BackgroundSelectionDialog/BackgroundSelectionDialog';
 import useChatContext from '../../hooks/useChatContext/useChatContext';
-import VideoLogo from '../IntroContainer/VideoLogo';
+import useVideoContext from '../../hooks/useVideoContext/useVideoContext';
 import ReactPlayer from 'react-player';
 import useParticipants from '../../hooks/useParticipants/useParticipants';
 import useMainParticipant from '../../hooks/useMainParticipant/useMainParticipant';
@@ -26,7 +27,7 @@ const useStyles = makeStyles((theme: Theme) => {
         gridTemplateRows: `calc(100% - ${totalMobileSidebarHeight}) ${totalMobileSidebarHeight}`,
       },
     },
-    chatWindowOpen: { gridTemplateColumns: `1fr ${theme.sidebarWidth}px ${theme.chatWindowWidth}px` },
+    rightDrawerOpen: { gridTemplateColumns: `1fr ${theme.sidebarWidth}px ${theme.rightDrawerWidth}px` },
   };
 });
 
@@ -43,11 +44,17 @@ function getQueryParams(queryString: string) {
 export default function Room() {
   const classes = useStyles();
   const { isChatWindowOpen } = useChatContext();
+  const { isBackgroundSelectionOpen } = useVideoContext();
+
   const participants = useParticipants();
   const { URLRoomName, URLPersona, URLName } = getQueryParams(window.location.search);
   if (participants.length === 0 && URLPersona !== 'provider')
     return (
-      <div className={clsx(classes.container, { [classes.chatWindowOpen]: isChatWindowOpen })}>
+      <div
+        className={clsx(classes.container, {
+          [classes.rightDrawerOpen]: isChatWindowOpen || isBackgroundSelectionOpen,
+        })}
+      >
         <ReactPlayer
           url="https://www.youtube.com/embed/E1h2Aqr8cu8"
           width="100%"
@@ -57,14 +64,17 @@ export default function Room() {
           loop={true}
           playing={true}
         />
-
         <ParticipantList />
         <ChatWindow />
       </div>
     );
   else
     return (
-      <div className={clsx(classes.container, { [classes.chatWindowOpen]: isChatWindowOpen })}>
+      <div
+        className={clsx(classes.container, {
+          [classes.rightDrawerOpen]: isChatWindowOpen || isBackgroundSelectionOpen,
+        })}
+      >
         <MainParticipant />
         <ParticipantList />
         <ChatWindow />